requirements.txt
amqp==2.2.2
autopep8==1.3.5
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-countries==5.3
django-debug-toolbar==1.9.1
django-debug-toolbar-template-profiler==1.0.2
django-guardian==1.4.9
django-js-asset==1.1.0
django-mptt==0.9.0
django-mptt-admin==0.5.1
django-npm==1.0.0
django-pagedown==1.0.5
django-redis==4.9.0
django-settings-export==1.2.1
django-silk==3.0.0
django-stdimage==3.1.1
djdt-flamegraph==0.2.12
gprof2dot==2016.10.13
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
markdown2==2.3.5
MarkupSafe==1.0
Pillow==5.1.0
progressbar2==3.37.1
pycodestyle==2.4.0
Pygments==2.2.0
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.4
redis==2.10.6
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
vine==1.1.4
